Put component definitions in another file
